composer install, php artisan serve, php artisan migrate:fresh seed

+++ database/migrations/2023_09_26_113708_create_products_table.php
@@ -7,23 +7,26 @@ use Illuminate\Support\Facades\Schema;
 return new class extends Migration
 {
     private string $table = 'products';
+
     /**
      * Run the migrations.
      */
     public function up(): void
     {
-        Schema::create($this->table, function (Blueprint $table) {
-            $table->id();
-            $table->string('name');
-            $table->text('short_description');
-            $table->text('long_description');
-            $table->foreignId('category_id')->constrained('product_categories')->onUpdate('cascade')->onDelete('cascade');;
-            $table->boolean('is_variable')->default(0);
-            $table->boolean('is_grouped')->default(0);
-            $table->boolean('is_simple')->default(1);
-            $table->string('featured_image');
-            $table->timestamps();
-        });
+        if (!Schema::hasTable($this->table)) {
+            Schema::create($this->table, function (Blueprint $table) {
+                $table->id();
+                $table->string('name');
+                $table->text('short_description');
+                $table->text('long_description');
+                $table->foreignId('category_id')->constrained('product_categories')->onUpdate('cascade')->onDelete('cascade');;
+                $table->boolean('is_variable')->default(0);
+                $table->boolean('is_grouped')->default(0);
+                $table->boolean('is_simple')->default(1);
+                $table->string('featured_image');
+                $table->timestamps();
+            });
+        }
     }
 
     /**
